Not quite sure I didn't make a mess in SofaPython.

Conflicts:
	applications/plugins/SofaPython/CMakeLists.txt
	applications/plugins/SofaPython/PythonCommon.h
	applications/plugins/SofaPython/PythonEnvironment.cpp
	extlibs/qwt-6.0.1/src/CMakeLists.txt
	framework/framework_test/CMakeLists.txt
	framework/sofa/helper/system/config.h
	scripts/ci/configure.sh

Former-commit-id: b15048ad19de9ca6f9c4bc6e83b5d5734c72dde3

+++ modules/SofaNonUniformFem/MultilevelHexahedronSetTopologyContainer.cpp
@@ -69,7 +69,7 @@ void MultilevelHexahedronSetTopologyContainer::init()
 
     _fineComponentInRegularGrid.resize(_fineResolution[0] * _fineResolution[1] * _fineResolution[2], NULL);
 
-    const unsigned int numVoxels = d_hexahedron.getValue().size();
+    const size_t numVoxels = d_hexahedron.getValue().size();
 
     // initialize the components
     // at the beginning the components of both levels are the same
